# Conflicts:
#	FF1Blazorizer/wwwroot/tooltips/tooltips.json
#	FF1Lib/FlagsViewModel.cs

+++ FF1Lib/EnemyFormations.cs
@@ -228,7 +228,7 @@ namespace FF1Lib
 			{
 				if (IsBossTrapTile(tile))
 				{
-					tile[1] = extendedtraptiles ? 0x00 : 0x80;
+					tile[1] = (byte)(extendedtraptiles ? 0x00 : 0x80);
 				}
 			});
 			Put(TilesetDataOffset, tilesets.SelectMany(tileset => tileset.ToBytes()).ToArray());
